Added .changes file to OBS build.
The .changes file is the openSUSE approach to keeping %changelog
up to date in the .spec files. For more info see:
http://en.opensuse.org/openSUSE:Specfile_guidelines

The recommended way of updating the .changes file is via the osc
command, as follows:
osc vc obs/qtgpsc.changes
(or just "osc vc" if inside the obs/ directory).